Added pwm generator with a gamma correction table.
This example fades two LEDs on the iCEBreaker board but uses a gamma
correction lookup table. This makes the LEDs fade more evenly. Our
eyesight is not linear but exponential, this means that if we fade the
LEDs linearelly they will go from off to on very quickly and seem to be
at a high brightness level for a very long time. The gamma correction
makes it seem that we fade the led value more linearelly. :)