Create the strobe on posedge not negedge.
I originally chose negedge as I thought I have to make sure the signal
is stable on the next rising edge of the clock. But thanks to @daveshah1
I now know that it is not necessary and the synthesis and routing tools
are there to take care of issues like this. Additionally having a mix of
posedge and negedge can cause worse timing performance as it can
decrease the frequency of the critical path by half.