Removed support arachne-pnr.
Arachne-pnr is depricated and superseded by nextpnr. If there is
some very good reason why you need arachne-pnr, like for example nextpnr
is not supported on your platform or you need some specific feature of
arachne-pnr. Please consider opening an issue on nextpnr so that your
problem can be addressed.