bitsy0 pcf: Change UART pins
Theses are the two pins that have been used for the
bootloader and other platform files until now ...

They are on the "programming" connector side.
The current pins (47/44) are not next to each other
and in the middle of one of the side, not great so
this is probably better and doesn't matter much since
it's v0 anyway.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>